# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WeTravel API
  description: Python backend with Clerk sessions and Supabase. Authorize with a session
    JWT, never a secret key.
  version: 2.0.0
paths:
...
  /api/media/upload:
    post:
      tags:
      - Media
      summary: Create Upload
      description: Return a direct upload URL/token. JPEG, PNG, WebP only; max 10
        MiB in Storage.
      operationId: create_upload_api_media_upload_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInput'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - ClerkSession: []
  /api/media/read:
    post:
      tags:
      - Media
      summary: Read Photo
      description: Signed URL valid for 300 seconds; ownership/publication checked
        by Storage RLS.
      operationId: read_photo_api_media_read_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadMediaInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadMediaResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - ClerkSession: []
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
        issues:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Issues
      type: object
      required:
      - error
      title: ErrorResponse
...
    ReadMediaInput:
      properties:
        path:
          type: string
          maxLength: 400
          pattern: ^[\w/-]+\.(jpg|png|webp)$
          title: Path
      additionalProperties: false
      type: object
      required:
      - path
      title: ReadMediaInput
    ReadMediaResponse:
      properties:
        signedUrl:
          type: string
          title: Signedurl
      type: object
      required:
      - signedUrl
      title: ReadMediaResponse
...
    UploadInput:
      properties:
        extension:
          type: string
          enum:
          - jpg
          - png
          - webp
          title: Extension
      additionalProperties: false
      type: object
      required:
      - extension
      title: UploadInput
    UploadResponse:
      properties:
        path:
          type: string
          title: Path
        signedUrl:
          type: string
          title: Signedurl
        token:
          type: string
          title: Token
      type: object
      required:
      - path
      - signedUrl
      - token
      title: UploadResponse
...
  securitySchemes:
    ClerkSession:
      type: http
      description: Clerk session JWT from getToken(); never your secret key.
      scheme: bearer